refactor: Fix test data
